Refactor the code and remove unused code
