+++ shoot/schema_gcp.go
@@ -1,152 +1,106 @@
 package shoot
 
-import (
-	"github.com/hashicorp/terraform/helper/schema"
-)
+import "github.com/hashicorp/terraform/helper/schema"
 
-func cloudResourceGCP() *schema.Resource {
+func gcpControlPlaneResource() *schema.Resource {
 	return &schema.Resource{
 		Schema: map[string]*schema.Schema{
-			"machine_image": {
-				Type:             schema.TypeList,
-				Description:      "MachineImage holds information about the machine image to use for all workers.",
-				Optional:         true,
-				MaxItems:         1,
-				DiffSuppressFunc: suppressMissingOptionalConfigurationBlock,
-				Elem: &schema.Resource{
-					Schema: map[string]*schema.Schema{
-						"name": {
-							Type:        schema.TypeString,
-							Description: "Name is the name of the image.",
-							Optional:    true,
-							Default:     "coreos",
-						},
-						"version": {
-							Type:             schema.TypeString,
-							Description:      "Version is the version of the image.",
-							Optional:         true,
-							DiffSuppressFunc: suppressEmptyNewValue,
-						},
-					},
-				},
+			"zone": {
+				Type:        schema.TypeString,
+				Description: "Zone is the GCP zone.",
+				Required:    true,
 			},
+		},
+	}
+}
+
+func gcpResource() *schema.Resource {
+	return &schema.Resource{
+		Schema: map[string]*schema.Schema{
 			"networks": {
 				Type:        schema.TypeList,
-				Description: "Networks holds information about the Kubernetes and infrastructure networks.",
+				Description: "Networks is the network configuration (VPC, subnets, etc.)",
 				Required:    true,
-				MaxItems:    1,
 				Elem: &schema.Resource{
 					Schema: map[string]*schema.Schema{
+						"vpc": {
+							Type:        schema.TypeList,
+							Description: "VPC indicates whether to use an existing VPC or create a new one.",
+							Optional:    true,
+							Elem: &schema.Resource{
+								Schema: map[string]*schema.Schema{
+									"name": {
+										Type:        schema.TypeString,
+										Description: "Name is the VPC name.",
+										Optional:    true,
+									},
+									"cloud_router": {
+										Type:        schema.TypeList,
+										Description: "CloudRouter indicates whether to use an existing CloudRouter or create a new one.",
+										Optional:    true,
+										Elem: &schema.Resource{
+											Schema: map[string]*schema.Schema{
+												"name": {
+													Type:        schema.TypeString,
+													Description: "Name is the CloudRouter name.",
+													Optional:    true,
+												},
+											},
+										},
+									},
+								},
+							},
+						},
+						"cloud_nat": {
+							Type:        schema.TypeList,
+							Description: "CloudNAT contains configuration about the the CloudNAT configuration",
+							Optional:    true,
+							Elem: &schema.Resource{
+								Schema: map[string]*schema.Schema{
+									"min_ports_per_vm": {
+										Type:        schema.TypeInt,
+										Description: "MinPortsPerVM is the minimum number of ports allocated to a VM in the NAT config. The default value is 2048 ports.",
+										Optional:    true,
+									},
+								},
+							},
+						},
 						"internal": {
 							Type:        schema.TypeString,
 							Description: "Internal is a private subnet (used for internal load balancers).",
 							Optional:    true,
 						},
 						"workers": {
-							Type:        schema.TypeSet,
-							Description: "Workers is a list of worker subnets (private) to create (used for the VMs).",
-							Required:    true,
-							Elem:        &schema.Schema{Type: schema.TypeString},
-							Set:         schema.HashString,
-						},
-					},
-				},
-			},
-			"worker": {
-				Type:        schema.TypeList,
-				Description: "Workers is a list of worker groups.",
-				Required:    true,
-				Elem: &schema.Resource{
-					Schema: map[string]*schema.Schema{
-						"name": {
 							Type:        schema.TypeString,
-							Description: "Name is the name of the worker group.",
+							Description: "Workers is the worker subnet range to create (used for the VMs).",
 							Required:    true,
 						},
-						"machine_type": {
-							Type:        schema.TypeString,
-							Description: "MachineType is the machine type of the worker group.",
-							Required:    true,
-						},
-						"auto_scaler_min": {
-							Type:        schema.TypeInt,
-							Description: "AutoScalerMin is the minimum number of VMs to create.",
-							Required:    true,
-						},
-						"auto_scaler_max": {
-							Type:        schema.TypeInt,
-							Description: "AutoScalerMax is the maximum number of VMs to create.",
-							Required:    true,
-						},
-						"max_surge": {
-							Type:        schema.TypeInt,
-							Description: "MaxSurge is maximum number of VMs that are created during an update.",
-							Optional:    true,
-						},
-						"max_unavailable": {
-							Type:        schema.TypeInt,
-							Description: "MaxUnavailable is the maximum number of VMs that can be unavailable during an update.",
-							Optional:    true,
-						},
-						"annotations": {
-							Type:        schema.TypeMap,
-							Description: "Annotations is a map of key/value pairs for annotations for all the `Node` objects in this worker pool.",
-							Optional:    true,
-						},
-						"labels": {
-							Type:        schema.TypeMap,
-							Description: "Labels is a map of key/value pairs for labels for all the `Node` objects in this worker pool.",
-							Optional:    true,
-						},
-						"taints": {
+						"flow_logs": {
 							Type:        schema.TypeList,
-							Description: "Taints is a list of taints for all the `Node` objects in this worker pool.",
+							Description: "FlowLogs contains the flow log configuration for the subnet.",
 							Optional:    true,
 							Elem: &schema.Resource{
 								Schema: map[string]*schema.Schema{
-									"key": {
-										Type:     schema.TypeString,
-										Required: true,
+									"aggregation_interval": {
+										Type:        schema.TypeString,
+										Description: "AggregationInterval for collecting flow logs.",
+										Optional:    true,
 									},
-									"operator": {
-										Type:     schema.TypeString,
-										Required: true,
+									"flow_sampling": {
+										Type:        schema.TypeFloat,
+										Description: "FlowSampling sets the sampling rate of VPC flow logs within the subnetwork where 1.0 means all collected logs are reported and 0.0 means no logs are reported.",
+										Optional:    true,
 									},
-									"value": {
-										Type:     schema.TypeString,
-										Optional: true,
-									},
-									"effect": {
-										Type:     schema.TypeString,
-										Required: true,
+									"metadata": {
+										Type:        schema.TypeString,
+										Description: "Metadata configures whether metadata fields should be added to the reported VPC flow logs.",
+										Optional:    true,
 									},
 								},
 							},
 						},
-						"volume_type": {
-							Type:        schema.TypeString,
-							Description: "VolumeType is the type of the root volumes.",
-							Required:    true,
-						},
-						"volume_size": {
-							Type:        schema.TypeString,
-							Description: "VolumeSize is the size of the root volume.",
-							Required:    true,
-						},
 					},
 				},
-			},
-			"zones": {
-				Type:        schema.TypeSet,
-				Description: "Zones is a list of availability zones to deploy the Shoot cluster to.",
-				Required:    true,
-				Elem:        &schema.Schema{Type: schema.TypeString},
-				Set:         schema.HashString,
-			},
-			"cabundle": {
-				Type:        schema.TypeString,
-				Description: "caBundle configuration",
-				Optional:    true,
 			},
 		},
 	}
